pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hiring-assistant-chatbot"
version = "0.1.0"
description = "An intelligent hiring assistant chatbot for initial candidate screening, built with Streamlit and powered by an LLM."
readme = { file = "README.md", content-type = "text/markdown" }
license = { text = "MIT" }
requires-python = ">=3.10,<3.13"
authors = [
    { name = "Vipin Kumar", email = "krvipin15@tutamail.com" },
]
keywords = ["chatbot", "hiring", "recruitment", "llm", "streamlit", "ai"]
dependencies = [
    "cryptography>=45.0.6",
    "email-validator>=2.3.0",
    "loguru>=0.7.3",
    "openai>=1.102.0",
    "phonenumbers>=9.0.12",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.0",
]

[project.scripts]
generate-key = "scripts.generate_key:main"
decrypt-db = "scripts.decrypt_database:main"
